requirements.txt
astroid==2.0
autopep8==1.3.5
click==6.7
Flask==1.0.2
gherkin-official==4.1.3
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
pkg-resources==0.0.0
pycodestyle==2.4.0
pylint==2.0.0
six==1.11.0
typed-ast==1.1.0
Werkzeug==0.14.1
wrapt==1.10.11
